technical_test_back/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
pygraphviz==1.11
pytz==2023.3
sqlparse==0.4.4
